refactor: split editing_normal_mode into submodules
