+++ core/src/state/notebook/inner_state/editing_normal_mode/change2.rs
@@ -1,0 +1,73 @@
+use super::VimNormalState;
+use {
+    super::VimVisualState,
+    crate::{
+        Error, Event, KeyEvent, NotebookEvent, NumKey, Result,
+        db::CoreBackend,
+        state::notebook::{InnerState, NoteTreeState, NotebookState, directory, note, tabs},
+        transition::{
+            NormalModeTransition, NotebookTransition, VimKeymapKind, VisualModeTransition,
+        },
+    },
+};
+
+pub async fn consume(
+    state: &mut NotebookState,
+    n1: usize,
+    n2: usize,
+    event: Event,
+) -> Result<NotebookTransition> {
+    use Event::*;
+    use NormalModeTransition::*;
+
+    match event {
+        Key(KeyEvent::Num(n)) => {
+            let n2 = n + n2.saturating_mul(10);
+            state.inner_state =
+                InnerState::EditingNormalMode(super::VimNormalState::Change2(n1, n2));
+
+            Ok(NotebookTransition::None)
+        }
+        Key(KeyEvent::I) => {
+            let n = n1.saturating_mul(n2);
+            state.inner_state =
+                InnerState::EditingNormalMode(super::VimNormalState::ChangeInside(n));
+
+            ChangeInsideMode.into()
+        }
+        Key(KeyEvent::C) => {
+            let n = n1.saturating_mul(n2);
+            state.inner_state = InnerState::EditingInsertMode;
+            DeleteLinesAndInsert(n).into()
+        }
+        Key(KeyEvent::E | KeyEvent::W) => {
+            let n = n1.saturating_mul(n2);
+            state.inner_state = InnerState::EditingInsertMode;
+            DeleteWordEnd(n).into()
+        }
+        Key(KeyEvent::B) => {
+            let n = n1.saturating_mul(n2);
+            state.inner_state = InnerState::EditingInsertMode;
+            DeleteWordBack(n).into()
+        }
+        Key(KeyEvent::DollarSign) => {
+            let n = n1.saturating_mul(n2);
+            state.inner_state = InnerState::EditingInsertMode;
+            DeleteLineEnd(n).into()
+        }
+        Key(KeyEvent::CtrlH) => Ok(NotebookTransition::ShowVimKeymap(
+            VimKeymapKind::NormalChange2,
+        )),
+        Key(KeyEvent::Esc) => {
+            state.inner_state = InnerState::EditingNormalMode(super::VimNormalState::Idle);
+
+            IdleMode.into()
+        }
+        event @ Key(_) => {
+            state.inner_state = InnerState::EditingNormalMode(super::VimNormalState::Idle);
+
+            super::idle::consume(state, event).await
+        }
+        _ => Err(Error::Wip("todo: Notebook::consume".to_owned())),
+    }
+}
