+++ core/src/state/notebook/inner_state/editing_normal_mode/change_inside.rs
@@ -1,0 +1,35 @@
+use super::VimNormalState;
+use {
+    super::VimVisualState,
+    crate::{
+        Error, Event, KeyEvent, NotebookEvent, NumKey, Result,
+        db::CoreBackend,
+        state::notebook::{InnerState, NoteTreeState, NotebookState, directory, note, tabs},
+        transition::{
+            NormalModeTransition, NotebookTransition, VimKeymapKind, VisualModeTransition,
+        },
+    },
+};
+
+pub async fn consume(
+    state: &mut NotebookState,
+    n: usize,
+    event: Event,
+) -> Result<NotebookTransition> {
+    use Event::*;
+    use NormalModeTransition::*;
+
+    match event {
+        Key(KeyEvent::W) => {
+            state.inner_state = InnerState::EditingInsertMode;
+
+            DeleteInsideWord(n).into()
+        }
+        event @ Key(_) => {
+            state.inner_state = InnerState::EditingNormalMode(super::VimNormalState::Idle);
+
+            super::idle::consume(state, event).await
+        }
+        _ => Err(Error::Wip("todo: Notebook::consume".to_owned())),
+    }
+}
