+++ core/src/state/notebook/inner_state/editing_normal_mode/gateway.rs
@@ -1,0 +1,36 @@
+use super::VimNormalState;
+use {
+    super::VimVisualState,
+    crate::{
+        Error, Event, KeyEvent, NotebookEvent, NumKey, Result,
+        db::CoreBackend,
+        state::notebook::{InnerState, NoteTreeState, NotebookState, directory, note, tabs},
+        transition::{
+            NormalModeTransition, NotebookTransition, VimKeymapKind, VisualModeTransition,
+        },
+    },
+};
+
+pub async fn consume(state: &mut NotebookState, event: Event) -> Result<NotebookTransition> {
+    use Event::*;
+    use NormalModeTransition::*;
+
+    match event {
+        Key(KeyEvent::G) => {
+            state.inner_state = InnerState::EditingNormalMode(VimNormalState::Idle);
+
+            NormalModeTransition::MoveCursorTop.into()
+        }
+        Key(KeyEvent::Esc) => {
+            state.inner_state = InnerState::EditingNormalMode(VimNormalState::Idle);
+
+            IdleMode.into()
+        }
+        event @ Key(_) => {
+            state.inner_state = InnerState::EditingNormalMode(VimNormalState::Idle);
+
+            super::idle::consume(state, event).await
+        }
+        _ => Err(Error::Wip("todo: Notebook::consume".to_owned())),
+    }
+}
